# backend/api_spec.yaml
openapi: "3.0.2"
info:
  title: "WaterProbe API"
  description: "API for waterprobing"
  version: "1"
paths:
  /Things:
    get:
      tags: 
      - Thing
      summary: "Gets all things (probes)"
      operationId: getThings
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
  /Thing:
    get:
      tags: 
      - Thing
      summary: "Get an existing probe"
      operationId: getThing
      parameters:
       - $ref: '#/components/parameters/thingname'
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
    post:
      tags:
       - Thing
      summary: "Creates a new probe"
      description: "Creates new thing"
      operationId: "NewThing"
      responses:
        201:
          $ref: '#/components/responses/201'
        404:
          $ref: '#/components/responses/404'
      requestBody:
        description: Optional properties represented as json for the Thing
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/thing'
              
    delete:
      tags: 
      - Thing
      summary: "Delete an existing probe"
      operationId: deleteThing
      parameters:
       - $ref: '#/components/parameters/thingname'
        
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
  /Observations:
    get:
      tags:
      - Observation
      summary: "Get all observations associated to datastream"
      operationId: getObservations
      responses:
        201:
          $ref: '#/components/responses/201'
        404:
          $ref: '#/components/responses/404'
      
      parameters: 
       - $ref: '#/components/parameters/datastreamRef'
  /Observation:
    get:
      tags:
      - Observation
      summary: "Call to get observation from database"
      operationId: getObservation
      responses:
        201:
          $ref: '#/components/responses/201'
        404:
          $ref: '#/components/responses/404'
      
      parameters: 
       - $ref: '#/components/parameters/observation'
    post:
      tags:
      - Observation
      summary: "Call to write observation to database"
      operationId: postObservation
      responses:
        201:
          $ref: '#/components/responses/201'
        404:
          $ref: '#/components/responses/404'
      
      requestBody:
        description: Optional properties represented as json for the Thing
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/observation'
...
  /dataStream:
    get:
      tags: 
      - Datastream
      summary: "Gets a datastream"
      operationId: getDatastream
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
                
      parameters: 
        - $ref: '#/components/parameters/dataStream'
        
    put:
      tags: 
      - Datastream
      summary: "put to a datastream"
      operationId: putDatastream
      responses:
        201:
          $ref: '#/components/responses/201'
        404:
          $ref: '#/components/responses/404'
                
      parameters: 
        - $ref: '#/components/parameters/dataStream'
        
      requestBody:
        description: Optional properties represented as json for the dataStream
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dataStream'
    
    post:
      tags: 
      - Datastream
      summary: "posts a datastream"
      operationId: postDatastream
      responses:
        201:
          $ref: '#/components/responses/201'
        404:
          $ref: '#/components/responses/404'
                
      requestBody:
        description: Optional properties represented as json for the dataStream
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dataStream'
    
    delete:
      tags: 
      - Datastream
      summary: "deletes a datastream"
      operationId: deleteDatastream
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
                
      parameters: 
        - $ref: '#/components/parameters/dataStream'
...
components:
  responses:
    200:
      description: Successful response
      content:
        application/json:
          schema:
            title: Sample
            type: object
            properties:
              placeholder:
                type: string
                description: Placeholder description
    201:
      description: Successful response
      content:
        application/json:
          schema:
            title: Sample
            type: object
            properties:
              placeholder:
                type: string
                description: Placeholder description
    404:
      description: Not created response
      content:
        text/plain:
          schema:
            title: Thing not cretead
            type: string
            example:
  parameters:
    observation: 
      name: observationName
      in: query
      description: "Unique observationName"
      required: false
      schema:
        type: string
    dataStream: 
      name: dataStreamName
      in: query
      description: "Unique dataStreamName"
      schema:
        type: string
...
    thingname: 
      name: thingname
      in: query
      description: "Unique thingName"
      schema:
        type: string
...
    datastreamRef:
      name: datastreamRef
      in: query
      description: "Ref to unique datastream"
      schema:
        type: string
  schemas:
...
    tmObject:
      type: object
      properties:
        time:
          type: string 
    dqElement:
      type: object
    tmInstant:
      type: object
      properties:
        time:
          type: string
    tmPeriod:
      type: object
    namedValue:
      type: object
    valueCode:
      type: object
    gm_envelope:
      type: object
    thing:
      type: object
      required:
        - name
        - description
        
      properties:
        name:
          type: string
        description: 
          type: string
        properties:
          type: array
          items:
            type: object
    observation:
      type: object
      required:
        - phenomenonTime 
        - resultTime
        - result
        - datastreamRef
        
      properties:
        phenomenonTime:
          $ref: '#/components/schemas/tmObject'
        resultTime:
          $ref: '#/components/schemas/tmInstant'
        result:
          type: number
        resultQuality:
          type: array
          items: 
            $ref: '#/components/schemas/dqElement'
        validTime:
          type: array
          items:
            $ref: '#/components/schemas/tmPeriod'
        parameters:
          type: array
          items:
            $ref: '#/components/schemas/namedValue'
        datastreamRef:
          type: string
        featureOfInterestRef:
          type: string
    dataStream:
      type: object
      required:
       - name
       - description
       - observationType
       - unitOfMeasurement
       - thingRef
       - sensorRef
       - observedPropertyRef
       
      properties:
        name:
          type: string
        description:
          type: string
        observationType:
          $ref: '#/components/schemas/valueCode'
        unitOfMeasurement:
          type: array	
          items:
            type: object
        observedArea:
          $ref: '#/components/schemas/gm_envelope'
        phenomenonTime:
          $ref: '#/components/schemas/tmPeriod'
        resultTime:
          $ref: '#/components/schemas/tmPeriod'
        thingRef:
          type: string
        sensorRef:
          type: string
        observedPropertyRef:
          type: string
...